Added application serving with flask and reorganized file structure to be compatible with flask library
