+++ Deprecated/c_algorithm.c
