+++ Deprecated/cpp_algorithm.cpp
