+++ Deprecated/java_algorithm.java
