+++ dependencies.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Upgrade pip
+pip install --upgrade pip
+
+# Install flask
+pip install flask
