--- index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <link rel="stylesheet" type="text/css" href="styles.css">
-    <script src="https://d3js.org/d3.v5.min.js"></script>
-</head>
-<body>
-    <svg id="canvas" width="800" height="600"></svg>
-    <script src="script.js"></script>
-</body>
-</html>
