--- styles.css
@@ -1,14 +1,0 @@
-body {
-    background-color: #f2f2f2;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-}
-
-#canvas {
-    border: 1px solid #ccc;
-    background-color: white;
-}
